+++ emeraldcomponents/src/main/java/br/com/stone/emeraldcomponents/basic/pinview/PinCodeCompleteListener.kt
@@ -1,0 +1,5 @@
+package br.com.stone.emeraldcomponents.basic.pinview
+
+interface PinCodeCompleteListener {
+    fun onCodeComplete(code: String)
+}