+++ emeraldcomponents/src/main/java/br/com/stone/emeraldcomponents/basic/pinview/PinCodeEventListener.kt
@@ -1,5 +1,0 @@
-package br.com.stone.emeraldcomponents.basic.pinview
-
-interface PinCodeEventListener {
-    fun onCodeFilled(code: String)
-}