+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## UNRELEASED
 
 ### Added
-- [Adds PinCode View][issue-170
+- [Adds PinCode View][issue-170]
 
 ## 1.11.0 - 2020-04-30