+++ CHANGELOG.md
@@ -239,3 +239,4 @@
 [issue-158]:https://github.com/stone-payments/emerald-components-android/issues/158
 [issue-162]:https://github.com/stone-payments/emerald-components-android/issues/162
 [issue-167]:https://github.com/stone-payments/emerald-components-android/issues/167
+[issue-170]:https://github.com/stone-payments/emerald-components-android/issues/170